# pkg/api/cluster/v1/cluster_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.8.0
// - protoc             (unknown)
// source: cluster/v1/cluster.proto

package clusterv1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationClusterServiceCreateCluster = "/cluster.v1.ClusterService/CreateCluster"
const OperationClusterServiceDeleteCluster = "/cluster.v1.ClusterService/DeleteCluster"
const OperationClusterServiceGetCluster = "/cluster.v1.ClusterService/GetCluster"
const OperationClusterServiceGetClusterHealth = "/cluster.v1.ClusterService/GetClusterHealth"
const OperationClusterServiceGetClusterVersion = "/cluster.v1.ClusterService/GetClusterVersion"
const OperationClusterServiceListClusters = "/cluster.v1.ClusterService/ListClusters"
const OperationClusterServiceUpdateCluster = "/cluster.v1.ClusterService/UpdateCluster"

type ClusterServiceHTTPServer interface {
	// CreateCluster 创建集群
	CreateCluster(context.Context, *CreateClusterRequest) (*Cluster, error)
	// DeleteCluster 删除集群
	DeleteCluster(context.Context, *DeleteClusterRequest) (*DeleteClusterResponse, error)
	// GetCluster 获取集群
	GetCluster(context.Context, *GetClusterRequest) (*Cluster, error)
	// GetClusterHealth 获取集群健康状态
	GetClusterHealth(context.Context, *GetClusterHealthRequest) (*ClusterHealth, error)
	// GetClusterVersion 获取集群版本信息
	GetClusterVersion(context.Context, *GetClusterVersionRequest) (*ClusterVersion, error)
	// ListClusters 列出集群
	ListClusters(context.Context, *ListClustersRequest) (*ListClustersResponse, error)
	// UpdateCluster 更新集群
	UpdateCluster(context.Context, *UpdateClusterRequest) (*Cluster, error)
}

func RegisterClusterServiceHTTPServer(s *http.Server, srv ClusterServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/api/v1/clusters/{cluster_id}", _ClusterService_GetCluster0_HTTP_Handler(srv))
	r.GET("/api/v1/clusters", _ClusterService_ListClusters0_HTTP_Handler(srv))
	r.POST("/api/v1/clusters", _ClusterService_CreateCluster0_HTTP_Handler(srv))
	r.PUT("/api/v1/clusters/{cluster_id}", _ClusterService_UpdateCluster0_HTTP_Handler(srv))
	r.DELETE("/api/v1/clusters/{cluster_id}", _ClusterService_DeleteCluster0_HTTP_Handler(srv))
	r.GET("/api/v1/clusters/{cluster_id}/health", _ClusterService_GetClusterHealth0_HTTP_Handler(srv))
	r.GET("/api/v1/clusters/{cluster_id}/version", _ClusterService_GetClusterVersion0_HTTP_Handler(srv))
}

func _ClusterService_GetCluster0_HTTP_Handler(srv ClusterServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetClusterRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationClusterServiceGetCluster)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetCluster(ctx, req.(*GetClusterRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Cluster)
		return ctx.Result(200, reply)
	}
}

func _ClusterService_ListClusters0_HTTP_Handler(srv ClusterServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListClustersRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationClusterServiceListClusters)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListClusters(ctx, req.(*ListClustersRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListClustersResponse)
		return ctx.Result(200, reply)
	}
}

func _ClusterService_CreateCluster0_HTTP_Handler(srv ClusterServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateClusterRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationClusterServiceCreateCluster)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateCluster(ctx, req.(*CreateClusterRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Cluster)
		return ctx.Result(200, reply)
	}
}

func _ClusterService_UpdateCluster0_HTTP_Handler(srv ClusterServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateClusterRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationClusterServiceUpdateCluster)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateCluster(ctx, req.(*UpdateClusterRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Cluster)
		return ctx.Result(200, reply)
	}
}

func _ClusterService_DeleteCluster0_HTTP_Handler(srv ClusterServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteClusterRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationClusterServiceDeleteCluster)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteCluster(ctx, req.(*DeleteClusterRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteClusterResponse)
		return ctx.Result(200, reply)
	}
}

func _ClusterService_GetClusterHealth0_HTTP_Handler(srv ClusterServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetClusterHealthRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationClusterServiceGetClusterHealth)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetClusterHealth(ctx, req.(*GetClusterHealthRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ClusterHealth)
		return ctx.Result(200, reply)
	}
}

func _ClusterService_GetClusterVersion0_HTTP_Handler(srv ClusterServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetClusterVersionRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationClusterServiceGetClusterVersion)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetClusterVersion(ctx, req.(*GetClusterVersionRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ClusterVersion)
		return ctx.Result(200, reply)
	}
}

type ClusterServiceHTTPClient interface {
	CreateCluster(ctx context.Context, req *CreateClusterRequest, opts ...http.CallOption) (rsp *Cluster, err error)
	DeleteCluster(ctx context.Context, req *DeleteClusterRequest, opts ...http.CallOption) (rsp *DeleteClusterResponse, err error)
	GetCluster(ctx context.Context, req *GetClusterRequest, opts ...http.CallOption) (rsp *Cluster, err error)
	GetClusterHealth(ctx context.Context, req *GetClusterHealthRequest, opts ...http.CallOption) (rsp *ClusterHealth, err error)
	GetClusterVersion(ctx context.Context, req *GetClusterVersionRequest, opts ...http.CallOption) (rsp *ClusterVersion, err error)
	ListClusters(ctx context.Context, req *ListClustersRequest, opts ...http.CallOption) (rsp *ListClustersResponse, err error)
	UpdateCluster(ctx context.Context, req *UpdateClusterRequest, opts ...http.CallOption) (rsp *Cluster, err error)
}

type ClusterServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewClusterServiceHTTPClient(client *http.Client) ClusterServiceHTTPClient {
	return &ClusterServiceHTTPClientImpl{client}
}

func (c *ClusterServiceHTTPClientImpl) CreateCluster(ctx context.Context, in *CreateClusterRequest, opts ...http.CallOption) (*Cluster, error) {
	var out Cluster
	pattern := "/api/v1/clusters"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationClusterServiceCreateCluster))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ClusterServiceHTTPClientImpl) DeleteCluster(ctx context.Context, in *DeleteClusterRequest, opts ...http.CallOption) (*DeleteClusterResponse, error) {
	var out DeleteClusterResponse
	pattern := "/api/v1/clusters/{cluster_id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationClusterServiceDeleteCluster))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ClusterServiceHTTPClientImpl) GetCluster(ctx context.Context, in *GetClusterRequest, opts ...http.CallOption) (*Cluster, error) {
	var out Cluster
	pattern := "/api/v1/clusters/{cluster_id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationClusterServiceGetCluster))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ClusterServiceHTTPClientImpl) GetClusterHealth(ctx context.Context, in *GetClusterHealthRequest, opts ...http.CallOption) (*ClusterHealth, error) {
	var out ClusterHealth
	pattern := "/api/v1/clusters/{cluster_id}/health"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationClusterServiceGetClusterHealth))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ClusterServiceHTTPClientImpl) GetClusterVersion(ctx context.Context, in *GetClusterVersionRequest, opts ...http.CallOption) (*ClusterVersion, error) {
	var out ClusterVersion
	pattern := "/api/v1/clusters/{cluster_id}/version"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationClusterServiceGetClusterVersion))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ClusterServiceHTTPClientImpl) ListClusters(ctx context.Context, in *ListClustersRequest, opts ...http.CallOption) (*ListClustersResponse, error) {
	var out ListClustersResponse
	pattern := "/api/v1/clusters"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationClusterServiceListClusters))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ClusterServiceHTTPClientImpl) UpdateCluster(ctx context.Context, in *UpdateClusterRequest, opts ...http.CallOption) (*Cluster, error) {
	var out Cluster
	pattern := "/api/v1/clusters/{cluster_id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationClusterServiceUpdateCluster))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

const OperationK8SResourceServiceCreateResource = "/cluster.v1.K8sResourceService/CreateResource"
const OperationK8SResourceServiceDeleteResource = "/cluster.v1.K8sResourceService/DeleteResource"
const OperationK8SResourceServiceGetResource = "/cluster.v1.K8sResourceService/GetResource"
const OperationK8SResourceServiceListResources = "/cluster.v1.K8sResourceService/ListResources"
const OperationK8SResourceServiceUpdateResource = "/cluster.v1.K8sResourceService/UpdateResource"

type K8SResourceServiceHTTPServer interface {
	// CreateResource 创建资源
	CreateResource(context.Context, *CreateResourceRequest) (*Resource, error)
	// DeleteResource 删除资源
	DeleteResource(context.Context, *DeleteResourceRequest) (*DeleteResourceResponse, error)
	// GetResource 获取资源
	GetResource(context.Context, *GetResourceRequest) (*Resource, error)
	// ListResources 列出资源
	ListResources(context.Context, *ListResourcesRequest) (*ListResourcesResponse, error)
	// UpdateResource 更新资源
	UpdateResource(context.Context, *UpdateResourceRequest) (*Resource, error)
}

func RegisterK8SResourceServiceHTTPServer(s *http.Server, srv K8SResourceServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/api/v1/clusters/{cluster_id}/resources/{resource_type}/{namespace}/{name}", _K8SResourceService_GetResource0_HTTP_Handler(srv))
	r.GET("/api/v1/clusters/{cluster_id}/resources/{resource_type}", _K8SResourceService_ListResources0_HTTP_Handler(srv))
	r.POST("/api/v1/clusters/{cluster_id}/resources/{resource_type}", _K8SResourceService_CreateResource0_HTTP_Handler(srv))
	r.PUT("/api/v1/clusters/{cluster_id}/resources/{resource_type}/{namespace}/{name}", _K8SResourceService_UpdateResource0_HTTP_Handler(srv))
	r.DELETE("/api/v1/clusters/{cluster_id}/resources/{resource_type}/{namespace}/{name}", _K8SResourceService_DeleteResource0_HTTP_Handler(srv))
}

func _K8SResourceService_GetResource0_HTTP_Handler(srv K8SResourceServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetResourceRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationK8SResourceServiceGetResource)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetResource(ctx, req.(*GetResourceRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Resource)
		return ctx.Result(200, reply)
	}
}

func _K8SResourceService_ListResources0_HTTP_Handler(srv K8SResourceServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListResourcesRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationK8SResourceServiceListResources)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListResources(ctx, req.(*ListResourcesRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListResourcesResponse)
		return ctx.Result(200, reply)
	}
}

func _K8SResourceService_CreateResource0_HTTP_Handler(srv K8SResourceServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateResourceRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationK8SResourceServiceCreateResource)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateResource(ctx, req.(*CreateResourceRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Resource)
		return ctx.Result(200, reply)
	}
}

func _K8SResourceService_UpdateResource0_HTTP_Handler(srv K8SResourceServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateResourceRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationK8SResourceServiceUpdateResource)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateResource(ctx, req.(*UpdateResourceRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Resource)
		return ctx.Result(200, reply)
	}
}

func _K8SResourceService_DeleteResource0_HTTP_Handler(srv K8SResourceServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteResourceRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationK8SResourceServiceDeleteResource)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteResource(ctx, req.(*DeleteResourceRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteResourceResponse)
		return ctx.Result(200, reply)
	}
}

type K8SResourceServiceHTTPClient interface {
	CreateResource(ctx context.Context, req *CreateResourceRequest, opts ...http.CallOption) (rsp *Resource, err error)
	DeleteResource(ctx context.Context, req *DeleteResourceRequest, opts ...http.CallOption) (rsp *DeleteResourceResponse, err error)
	GetResource(ctx context.Context, req *GetResourceRequest, opts ...http.CallOption) (rsp *Resource, err error)
	ListResources(ctx context.Context, req *ListResourcesRequest, opts ...http.CallOption) (rsp *ListResourcesResponse, err error)
	UpdateResource(ctx context.Context, req *UpdateResourceRequest, opts ...http.CallOption) (rsp *Resource, err error)
}

type K8SResourceServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewK8SResourceServiceHTTPClient(client *http.Client) K8SResourceServiceHTTPClient {
	return &K8SResourceServiceHTTPClientImpl{client}
}

func (c *K8SResourceServiceHTTPClientImpl) CreateResource(ctx context.Context, in *CreateResourceRequest, opts ...http.CallOption) (*Resource, error) {
	var out Resource
	pattern := "/api/v1/clusters/{cluster_id}/resources/{resource_type}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationK8SResourceServiceCreateResource))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *K8SResourceServiceHTTPClientImpl) DeleteResource(ctx context.Context, in *DeleteResourceRequest, opts ...http.CallOption) (*DeleteResourceResponse, error) {
	var out DeleteResourceResponse
	pattern := "/api/v1/clusters/{cluster_id}/resources/{resource_type}/{namespace}/{name}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationK8SResourceServiceDeleteResource))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *K8SResourceServiceHTTPClientImpl) GetResource(ctx context.Context, in *GetResourceRequest, opts ...http.CallOption) (*Resource, error) {
	var out Resource
	pattern := "/api/v1/clusters/{cluster_id}/resources/{resource_type}/{namespace}/{name}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationK8SResourceServiceGetResource))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *K8SResourceServiceHTTPClientImpl) ListResources(ctx context.Context, in *ListResourcesRequest, opts ...http.CallOption) (*ListResourcesResponse, error) {
	var out ListResourcesResponse
	pattern := "/api/v1/clusters/{cluster_id}/resources/{resource_type}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationK8SResourceServiceListResources))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *K8SResourceServiceHTTPClientImpl) UpdateResource(ctx context.Context, in *UpdateResourceRequest, opts ...http.CallOption) (*Resource, error) {
	var out Resource
	pattern := "/api/v1/clusters/{cluster_id}/resources/{resource_type}/{namespace}/{name}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationK8SResourceServiceUpdateResource))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}