# pkg/api/docs/swagger/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "common/pagination/v1/pagination.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/clusters": {
      "get": {
        "summary": "列出集群",
        "operationId": "ClusterService_ListClusters",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListClustersResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "pagination.page",
            "description": "页码（从 1 开始）",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pagination.pageSize",
            "description": "每页大小",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pagination.orderBy",
            "description": "排序字段",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "pagination.orderDirection",
            "description": "排序方向：asc/desc",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "status",
            "description": " - HEALTHY: 健康\n - UNHEALTHY: 不健康\n - UNREACHABLE: 不可达\n - UNKNOWN: 未知",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "CLUSTER_STATUS_UNSPECIFIED",
              "HEALTHY",
              "UNHEALTHY",
              "UNREACHABLE",
              "UNKNOWN"
            ],
            "default": "CLUSTER_STATUS_UNSPECIFIED"
          }
        ],
        "tags": [
          "ClusterService"
        ]
      },
      "post": {
        "summary": "创建集群",
        "operationId": "ClusterService_CreateCluster",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Cluster"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateClusterRequest"
            }
          }
        ],
        "tags": [
          "ClusterService"
        ]
      }
    },
    "/api/v1/clusters/{clusterId}": {
      "get": {
        "summary": "获取集群",
        "operationId": "ClusterService_GetCluster",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Cluster"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "clusterId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "ClusterService"
        ]
      },
      "delete": {
        "summary": "删除集群",
        "operationId": "ClusterService_DeleteCluster",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteClusterResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "clusterId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "ClusterService"
        ]
      },
      "put": {
        "summary": "更新集群",
        "operationId": "ClusterService_UpdateCluster",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Cluster"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "clusterId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ClusterServiceUpdateClusterBody"
            }
          }
        ],
        "tags": [
          "ClusterService"
        ]
      }
    },
...
    "/api/v1/clusters/{clusterId}/resources/{resourceType}": {
      "get": {
        "summary": "列出资源",
        "operationId": "K8sResourceService_ListResources",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListResourcesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "clusterId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "resourceType",
            "in": "path",
            "required": true,
            "type": "string",
            "enum": [
              "RESOURCE_TYPE_UNSPECIFIED",
              "POD",
              "DEPLOYMENT",
              "SERVICE",
              "CONFIGMAP",
              "SECRET",
              "NAMESPACE",
              "NODE",
              "PERSISTENTVOLUME",
              "PERSISTENTVOLUMECLAIM",
              "STATEFULSET",
              "DAEMONSET",
              "JOB",
              "CRONJOB",
              "INGRESS",
              "NETWORKPOLICY",
              "RESOURCEQUOTA",
              "LIMITRANGE",
              "HORIZONTALPODAUTOSCALER",
              "STORAGECLASS",
              "ROLE",
              "ROLEBINDING",
              "CLUSTERROLE",
              "CLUSTERROLEBINDING",
              "SERVICEACCOUNT",
              "ENDPOINT",
              "ENDPOINTSLICE",
              "EVENT",
              "PRIORITYCLASS",
              "REPLICASET"
            ]
          },
          {
            "name": "namespace",
            "description": "空表示所有命名空间",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "labelSelector",
            "description": "标签选择器",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "pagination.page",
            "description": "页码（从 1 开始）",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pagination.pageSize",
            "description": "每页大小",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pagination.orderBy",
            "description": "排序字段",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "pagination.orderDirection",
            "description": "排序方向：asc/desc",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "K8sResourceService"
        ]
      },
      "post": {
        "summary": "创建资源",
        "operationId": "K8sResourceService_CreateResource",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Resource"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "clusterId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "resourceType",
            "in": "path",
            "required": true,
            "type": "string",
            "enum": [
              "RESOURCE_TYPE_UNSPECIFIED",
              "POD",
              "DEPLOYMENT",
              "SERVICE",
              "CONFIGMAP",
              "SECRET",
              "NAMESPACE",
              "NODE",
              "PERSISTENTVOLUME",
              "PERSISTENTVOLUMECLAIM",
              "STATEFULSET",
              "DAEMONSET",
              "JOB",
              "CRONJOB",
              "INGRESS",
              "NETWORKPOLICY",
              "RESOURCEQUOTA",
              "LIMITRANGE",
              "HORIZONTALPODAUTOSCALER",
              "STORAGECLASS",
              "ROLE",
              "ROLEBINDING",
              "CLUSTERROLE",
              "CLUSTERROLEBINDING",
              "SERVICEACCOUNT",
              "ENDPOINT",
              "ENDPOINTSLICE",
              "EVENT",
              "PRIORITYCLASS",
              "REPLICASET"
            ]
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/K8sResourceServiceCreateResourceBody"
            }
          }
        ],
        "tags": [
          "K8sResourceService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "ClusterServiceUpdateClusterBody": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "apiServer": {
          "type": "string"
        },
        "kubeconfig": {
          "type": "string"
        },
        "labels": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      },
      "title": "更新集群请求"
    },
...
    "K8sResourceServiceCreateResourceBody": {
      "type": "object",
      "properties": {
        "namespace": {
          "type": "string"
        },
        "spec": {
          "type": "object",
          "title": "资源规格（JSON格式）"
        }
      },
      "title": "创建资源请求"
    },
...
    "googlerpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1Cluster": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "apiServer": {
          "type": "string",
          "title": "API Server地址"
        },
        "kubeconfig": {
          "type": "string",
          "title": "Kubeconfig内容（base64编码）"
        },
        "version": {
          "type": "string",
          "title": "K8s版本"
        },
        "status": {
          "$ref": "#/definitions/v1ClusterStatus"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "labels": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      },
      "title": "集群信息"
    },
...
    "v1CreateClusterRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "apiServer": {
          "type": "string"
        },
        "kubeconfig": {
          "type": "string"
        },
        "labels": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      },
      "title": "创建集群请求"
    },
...
    "v1DeleteClusterResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "title": "删除集群响应"
    },
...
    "v1ListClustersResponse": {
      "type": "object",
      "properties": {
        "clusters": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Cluster"
          }
        },
        "pagination": {
          "$ref": "#/definitions/v1PaginationMetadata"
        }
      },
      "title": "列出集群响应"
    },
...
    "v1ListResourcesResponse": {
      "type": "object",
      "properties": {
        "resources": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Resource"
          }
        },
        "pagination": {
          "$ref": "#/definitions/v1PaginationMetadata"
        }
      },
      "title": "列出资源响应"
    },
...
    "v1Resource": {
      "type": "object",
      "properties": {
        "clusterId": {
          "type": "string"
        },
        "type": {
          "$ref": "#/definitions/v1ResourceType"
        },
        "namespace": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "spec": {
          "type": "object",
          "title": "资源规格（JSON格式）"
        },
        "status": {
          "type": "object",
          "title": "资源状态（JSON格式）"
        },
        "labels": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "annotations": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "title": "K8s资源"
    },
...
  }
}